Google-Gemini-Pro/requirements.txt
gTTS
numpy
opencv_python
Pillow
protobuf
python-dotenv
SpeechRecognition
streamlit
streamlit_webrtc
google-generativeai